Refactor PublicProfileModal to use React Query for data fetching and improve component structure with dedicated sections for profile details, badges, and activities.
